@@ -4,7 +4,7 @@
   "type": "craft-plugin",
   "license": "MIT",
   "require": {
-    "php": "^7.2.5",
+    "php": "^7.4.0",
     "ext-json": "*",
     "craftcms/cms": "^3.6.0",
     "sebastianlenz/craft-utils": "^2.1.0",
