@@ -4,13 +4,9 @@ namespace lenz\craft\essentials\structs\menu;
 
 use craft\base\ElementInterface;
 use craft\elements\Entry;
-use craft\helpers\Html;
-use craft\helpers\Template;
-use Exception;
 use lenz\craft\essentials\structs\structure\AbstractStructureItem;
-use Twig\Markup;
+use lenz\craft\utils\models\Attributes;
 use yii\base\InvalidConfigException;
-use yii\helpers\ArrayHelper;
 
 /**
  * Class AbstractMenuItem
@@ -18,83 +14,60 @@ use yii\helpers\ArrayHelper;
 abstract class AbstractMenuItem extends AbstractStructureItem
 {
   /**
-   * @var string|array
+   * @var array
    */
-  public $customLinkAttributes;
+  public array $customLinkAttributes;
 
   /**
    * @var bool
    */
-  public $isActive = false;
+  public bool $isActive = false;
 
   /**
    * @var string
    */
-  public $sectionHandle;
+  public string $sectionHandle;
 
   /**
    * @var int
    */
-  public $sectionId;
+  public int $sectionId;
 
   /**
    * @var string
    */
-  public $title;
+  public string $title;
 
   /**
    * @var string
    */
-  public $typeHandle;
+  public string $typeHandle;
 
   /**
    * @var int
    */
-  public $typeId;
+  public int $typeId;
 
   /**
    * @var string
    */
-  public $url;
+  public string $url;
 
 
   /**
-   * @param array $attributes
-   * @return Markup
+   * @param array $extraAttribs
+   * @return Attributes
    * @noinspection PhpUnused (Public/Template API)
-   * @throws Exception
    */
-  public function getLinkAttributes(array $attributes = []): Markup {
-    if (isset($this->customLinkAttributes)) {
-      if (is_string($this->customLinkAttributes)) {
-        if (!empty($attributes) && $_ENV['ENVIRONMENT'] !== 'production') {
-          throw new Exception('Cannot apply attributes when custom attributes are set to a string.');
-        }
-
-        return Template::raw($this->customLinkAttributes);
-      } else {
-        $attributes = array_merge($attributes, $this->customLinkAttributes);
-      }
-    } else {
-      $attributes = array_merge($attributes, [
-        'href' => $this->url
-      ]);
-    }
+  public function getLinkAttributes(array $extraAttribs = []): Attributes {
+    $attr = new Attributes($this->customLinkAttributes ?? ['href' => $this->url]);
+    $attr->set($extraAttribs);
 
     if ($this->isActive) {
-      $classNames = ArrayHelper::getValue($attributes, 'class', []);
-      if (!is_array($classNames)) {
-        $classNames = array_filter(preg_split('/\s+/', $classNames));
-      }
-
-      if (!in_array('active', $classNames)) {
-        $classNames[] = 'active';
-      }
-
-      $attributes['class'] = $classNames;
+      $attr->addClass('active');
     }
 
-    return Template::raw(Html::renderTagAttributes($attributes));
+    return $attr;
   }
 
 
@@ -107,17 +80,14 @@ abstract class AbstractMenuItem extends AbstractStructureItem
    */
   protected function setElement(ElementInterface $element) {
     parent::setElement($element);
-
-    if ($element instanceof ElementInterface) {
-      $this->title = (string)$element;
-      $this->url   = $element->getUrl();
-    }
+    $this->title = (string)$element;
+    $this->url = $element->getUrl();
 
     if ($element instanceof Entry) {
       $this->sectionHandle = $element->getSection()->handle;
-      $this->sectionId     = intval($element->sectionId);
-      $this->typeHandle    = $element->getType()->handle;
-      $this->typeId        = intval($element->typeId);
+      $this->sectionId = intval($element->sectionId);
+      $this->typeHandle = $element->getType()->handle;
+      $this->typeId = intval($element->typeId);
     }
   }
 }
